requirements.txt
fastapi==0.116.1
openai==1.99.9
pydantic==2.11.7
python-dotenv
python-multipart==0.0.20
uvicorn[standard]==0.35.0
